Fix Rubocop errors.
