+++ spec/spec_helper.rb
@@ -3,7 +3,7 @@ $LOAD_PATH.unshift(File.dirname(__FILE__))
 
 require 'rubygems'
 require 'rspec'
-require 'artsy-client'
+require 'artsy_client'
 
 require 'webmock/rspec'
 WebMock.disable_net_connect!
